feat: navbar dropdown on hover
